aqsa

+++ main.py
@@ -58,8 +58,8 @@ class DiabetesData(BaseModel):
     hypertension: int
     heart_disease: int
     smoking_history: int
-    bmi: int
-    HbA1c_level: int
+    bmi: float
+    HbA1c_level: float
     blood_glucose_level: int
 
 class ImageData(BaseModel):
